+++ src/controllers/VerifyController.php
@@ -10,6 +10,7 @@ use born05\twofactorauthentication\web\assets\verify\VerifyAsset;
 
 use yii\base\Event;
 use yii\web\UserEvent;
+use yii\web\Response;
 
 class VerifyController extends Controller
 {
@@ -46,22 +47,12 @@ class VerifyController extends Controller
 
             return $this->_handleSuccessfulLogin(true);
         } else {
-            $errorCode = User::AUTH_INVALID_CREDENTIALS;
-            $errorMessage = Craft::t('two-factor-authentication', 'Authentication code is invalid.');
-
-            if ($request->getAcceptsJson()) {
-                return $this->asJson([
-                    'errorCode' => $errorCode,
-                    'error' => $errorMessage
-                ]);
-            } else {
-                Craft::$app->getSession()->setError($errorMessage);
-
-                Craft::$app->getUrlManager()->setRouteParams([
-                    'errorCode' => $errorCode,
-                    'errorMessage' => $errorMessage,
-                ]);
-            }
+            return $this->asFailure(
+                Craft::t('two-factor-authentication', 'Authentication code is invalid.'),
+                data: [
+                    'errorCode' => User::AUTH_INVALID_CREDENTIALS,
+                ],
+            );
         }
     }
 
@@ -71,36 +62,23 @@ class VerifyController extends Controller
      * Redirects the user after a successful login attempt, or if they visited the Login page while they were already
      * logged in.
      *
-     * @param bool $setNotice Whether a flash notice should be set, if this isn't an Ajax request.
-     *
-     * @return null
+     * @return Response
      */
-    private function _handleSuccessfulLogin($setNotice)
+    private function _handleSuccessfulLogin(): Response
     {
-        $request = Craft::$app->getRequest();
-        $returnUrl = TwoFactorAuth::$plugin->response->getReturnUrl();
+        // Get the return URL
+        $userSession = Craft::$app->getUser();
+        $returnUrl = $userSession->getReturnUrl();
+
+        // Clear it out
+        $userSession->removeReturnUrl();
 
         // If this was an Ajax request, just return success:true
-        if ($request->getAcceptsJson()) {
-            $return = [
-                'success' => true,
-                'returnUrl' => $returnUrl,
-            ];
-
-            if (Craft::$app->getConfig()->getGeneral()->enableCsrfProtection) {
-                $return['csrfTokenValue'] = $request->getCsrfToken();
-            }
-
-            return $this->asJson($return);
+        if ($this->request->getAcceptsJson()) {
+            return $this->asSuccess(redirect: $returnUrl);
         }
 
-        if ($setNotice) {
-            Craft::$app->getSession()->setNotice(Craft::t('app', 'Logged in.'));
-        }
-
-        $user = Craft::$app->getUser()->getIdentity();
-
-        return $this->redirectToPostedUrl($user, $returnUrl);
+        return $this->redirectToPostedUrl($userSession->getIdentity(), $returnUrl);
     }
 
     /**