+++ src/services/Verify.php
@@ -145,7 +145,7 @@ class Verify extends Component
             $userRecord = $this->getUserRecord($user);
             $this->totp = TOTP::create($userRecord->secret);
             $this->totp->setLabel($user->email);
-            $this->totp->setIssuer(Craft::$app->getInfo()->name);
+            $this->totp->setIssuer(Craft::$app->getSystemName());
         }
 
         return $this->totp;